+++ src/pie-chart/segments.tsx
@@ -4,35 +4,29 @@ import React, { useMemo } from 'react';
 import { arc, PieArcDatum } from 'd3-shape';
 
 import { PieChartProps } from './interfaces';
-import { getDimensionsBySize } from './utils';
+import { Dimension } from './utils';
 import { InternalChartDatum } from './pie-chart';
-import { useVisualRefresh } from '../internal/hooks/use-visual-mode';
 import styles from './styles.css.js';
 import clsx from 'clsx';
 import { useInternalI18n } from '../internal/i18n/context';
 
 interface SegmentsProps<T> {
-  height: number;
-  fitHeight?: boolean;
   pieData: Array<PieArcDatum<InternalChartDatum<T>>>;
   highlightedSegment: T | null;
-  size: NonNullable<PieChartProps['size']>;
+  dimensions: Dimension;
   variant: PieChartProps['variant'];
   focusedSegmentRef: React.RefObject<SVGGElement>;
   popoverTrackRef: React.RefObject<SVGCircleElement>;
   segmentAriaRoleDescription?: string;
-
   onMouseDown: (datum: InternalChartDatum<T>) => void;
   onMouseOver: (datum: InternalChartDatum<T>) => void;
   onMouseOut: (event: React.MouseEvent<SVGElement>) => void;
 }
 
 export default function Segments<T extends PieChartProps.Datum>({
-  height,
-  fitHeight,
   pieData,
   highlightedSegment,
-  size,
+  dimensions,
   variant,
   focusedSegmentRef,
   popoverTrackRef,
@@ -42,10 +36,8 @@ export default function Segments<T extends PieChartProps.Datum>({
   onMouseOut,
 }: SegmentsProps<T>) {
   const i18n = useInternalI18n('pie-chart');
-  const isRefresh = useVisualRefresh();
 
   const { arcFactory, highlightedArcFactory } = useMemo(() => {
-    const dimensions = getDimensionsBySize({ size: fitHeight ? height : size, visualRefresh: isRefresh });
     const radius = dimensions.outerRadius;
     const innerRadius = variant === 'pie' ? 0 : dimensions.innerRadius;
     const cornerRadius = dimensions.cornerRadius || 0;
@@ -63,7 +55,7 @@ export default function Segments<T extends PieChartProps.Datum>({
       arcFactory,
       highlightedArcFactory,
     };
-  }, [size, fitHeight, height, variant, isRefresh]);
+  }, [dimensions, variant]);
 
   const centroid = useMemo(() => {
     for (const datum of pieData) {
